+++ Editor/ViewEffect/BlinkSpriteRenderer.cs
@@ -1,0 +1,10 @@
+using UnityEngine;
+
+namespace DT.UniUtils.View {
+  [RequireComponent(typeof(SpriteRenderer))]
+  public class BlinkSpriteRenderer : BlinkBehaviour<SpriteRenderer> {
+    void Update() {
+      this.component.WithAlpha(this.GetAlpha());
+    }
+  }
+}