+++ Editor/ViewEffect/BlinkCanvasRenderer.cs
@@ -4,7 +4,13 @@ namespace DT.UniUtils.View {
   [RequireComponent(typeof(CanvasRenderer))]
   public class BlinkCanvasRenderer : BlinkBehaviour<CanvasRenderer> {
     void Update() {
-      this.component.SetAlpha(this.GetAlpha());
+      if (this.blinking)
+        this.component.SetAlpha(this.GetAlpha());
+    }
+
+    public void StopBlink() {
+      base.StopBlink();
+      this.component.SetAlpha(1);
     }
   }
 }